+++ spring-batch-notion/src/main/java/org/springframework/batch/extensions/notion/package-info.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 /**
  * Spring Batch extension for Notion.
  */
