+++ spring-batch-notion/src/test/java/org/springframework/batch/extensions/notion/builder/package-info.java
@@ -15,6 +15,6 @@
  */
 
 @NullUnmarked
-package org.springframework.batch.extensions.notion.it;
+package org.springframework.batch.extensions.notion.builder;
 
 import org.jspecify.annotations.NullUnmarked;
