+++ spring-batch-notion/src/test/java/org/springframework/batch/extensions/notion/package-info.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 /**
  * Tests for Spring Batch Notion.
  */
